requirements.txt
pandas==2.2.3
sqlalchemy==2.0.44
snowflake-sqlalchemy==1.7.7
snowflake-connector-python==3.18.0
langchain==0.3.7
langchain-openai==0.3.5
openai>=1.58.1,<2.0.0
python-dotenv==1.0.1
requests==2.32.3
streamlit